+++ lib/semantic_release.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require_relative "semantic_release/configuration"
 require_relative "semantic_release/semver"
 require_relative "semantic_release/version"
 Dir[File.join(__dir__, "semantic_release", "updaters", "*.rb")].each { |file| require file }
@@ -7,15 +8,17 @@ Dir[File.join(__dir__, "semantic_release", "updaters", "*.rb")].each { |file| re
 module SemanticRelease
   class Error < StandardError; end
 
-  SEMVER_FILE = ".semver"
+  def self.configuration
+    @configuration ||= Configuration.new
+  end
 
   def self.init
     version = Semver.new
-    version.save(SEMVER_FILE)
+    version.save(configuration.semver_file)
   end
 
   def self.current_version
-    Semver.load(SEMVER_FILE).to_s
+    Semver.load(configuration.semver_file).to_s
   end
 
   def self.inc_major
@@ -34,7 +37,7 @@ module SemanticRelease
   end
 
   def self.increment(term)
-    version = Semver.load(SEMVER_FILE)
+    version = Semver.load(configuration.semver_file)
     version.increment(term)
     version.save
   end