+++ lib/semantic_release/rake_task.rb
@@ -5,9 +5,11 @@ require "semantic_release"
 
 module SemanticRelease
   class RakeTask < Rake::TaskLib
-    # rubocop:disable Metrics/MethodLength
+    # rubocop:disable Metrics/AbcSize, Metrics/MethodLength
     def initialize(name = :release)
       super()
+
+      yield(SemanticRelease.configuration) if block_given?
 
       namespace name do
         desc "Initialise version file"
@@ -36,6 +38,6 @@ module SemanticRelease
         end
       end
     end
-    # rubocop:enable Metrics/MethodLength
+    # rubocop:enable Metrics/AbcSize, Metrics/MethodLength
   end
 end