+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Add configuration options for semver file
 - Change rake task namespace from `semantic_release` to `release` and make configurable
 
 ## 1.0.1 (02 February 2025)